Bug: Add preprocess_syn_ADPUPA_obs.py
The new script changes ADPUPA pressure levels in the prepBUFR CSV files so that the mandatory levels are exact (e.g., 500.0 hPa instead of 500.1 hPa). This is necessary for upper-air verification with METplus b/c METplus does not interpolate obs vertically, so if the ADPUPA pressure levels do not perfectly match the verification pressure levels, then the obs will be omitted for verification. This appears to only be a problem with the synthetic obs (not the real_red obs).