Feat: Add confidence intervals to die-off and upper-air profiles
These confidence intervals use the same approach as that in MATS. Specifically, the confidence interval calculation uses the standard deviation of all the RMSEs from each model initialization time. This is not the most correct way of computing confidence intervals, but it is the best we can do since the standard deviation of the squared errors (i.e., (f-o)^2) cannot be computed from the partial sums provided by METplus. So the confidence intervals being plotted are really the confidence intervals for the mean RMSE, not the total RMSE.