Add ability to mask pressure levels beneath the surface
Upper-air verification and lower-atmosphere verification now include the
ability to mask pressure levels that lie beneath the surface. To do
this, a separate mask must be created for each pressure level and each
variable/pressure level pair must be a separate verification variable
(which results in A LOT of verification variables in the conf files).