Add ability to compute RHobT
Similar to MATS, RHobT is computed using the temperature from the
observations (in this case, the nature run) along with specific humidity
from the forecast runs.

RHobT is computed using a Python script (that relies on pygrib) as well
as a Rocoto workflow. The goal is to eventually verify these RHobT
values using MET.