Add plotting errors and print statements for better debugging
Plotting scripts will now throw an error if a MET output file that it is
expecting to find is empty. I debated whether this should be a warning
or an error, but ultimately decided to make it an error. I may change my
mind later.

Additional information messages were also added to the timeseries
plotting function to make it easier to debug issues.