Add check to see whether a p levels exists before plotting diffs
Code now explicitly checks to see whether a pressure level exists in the control run before trying to create a difference upper-air plot. It is guaranteed that the pressure level will exist in the experiment run, but it is not guaranteed that this will be the case in the control run.